requirements.txt
crawl4ai
langchain
langchain-community
langchain-openai
chromadb
openai
tiktoken
streamlit
python-dotenv
unstructured
